heron-server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.2.1
build==1.4.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.7
dnspython==2.8.0
docutils==0.22.4
ecdsa==0.19.2
email_validator==2.2.0
fastapi==0.135.1
h11==0.16.0
id==1.6.1
idna==3.11
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jeepney==0.9.0
keyring==25.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==11.0.2
nh3==0.3.4
packaging==26.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyproject_hooks==1.2.0
pytest==9.0.3
python-dotenv==1.2.2
python-multipart==0.0.20
PyJWT==2.8.0
readme_renderer==44.0
requests==2.33.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==15.0.0
rsa==4.9.1
SecretStorage==3.5.0
six==1.17.0
starlette==0.52.1
twine==6.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
